renamed settings to app_settings for consistancy removed collectionversion model because that was app specific starting to update docs
