+++ dataforms/app_settings.py
@@ -1,24 +1,5 @@
 """
-App Settings
-============
-
-This file should not be updated.  Instead, please put any additional field mappings
-in your project settings file.  That way any custom fields and widgets will be
-de-coupled from this app.
-
-:constant FIELD_MAPPINGS: a dictionary of form fields are available to be used.
-:constant ADMIN_SORT_JS: tuple that references the custom javascript for the django admin.
-:constant FIELD_TYPE_CHOICES: tuple that is used by the model as choices
-
-:constant BOOLEAN_FIELDS: an element of FIELD_MAPPINGS that should be interpreted
-	on save as only yes/no
-:constant SINGLE_CHOICE_FIELDS = field types that have choices, but only one selected
-	on save, for Select boxes or a single RadioSelect type
-:constant MULTI_CHOICE_FIELDS = field types that have choices and multiple
-	ones can be selected on save, for SelectMultiple boxes or CheckboxSelectMultiple
-:constant CHOICE_FIELDS: both SINGLE_CHOICE_FIELDS and MULTI_CHOICE_FIELDS  
-:constant SINGLE_NUMBER_FIELDS: For fields that need pseudo-foreign key storage
-:constant MULTI_NUMBER_FIELDS: For fields that need pseudo-many-to-many storage
+Dataforms App Settings
 """
 
 from django.conf import settings
@@ -47,8 +28,6 @@ FIELD_MAPPINGS.update( {
 	# Note Widget:  This is a way you can add sub headings to your forms.  See - dataforms.widgets.NoteWidget
 	'Note' : { 'class': 'django.forms.CharField', 'widget': 'dataforms.widgets.NoteWidget' },
 	
-	
-	
 	# FIXME: Remove After testing....
 	'USStateField' : { 'class': 'django.contrib.localflavor.us.forms.USStateField', 'widget' : 'django.forms.TextInput' },   
 } )
@@ -63,7 +42,7 @@ BOOLEAN_FIELDS = getattr(settings, "DATAFORMS_BOOLEAN_FIELDS", ()) + ('CheckboxI
 SINGLE_CHOICE_FIELDS = getattr(settings, "DATAFORMS_SINGLE_CHOICE_FIELDS", ()) + ('Select', 'RadioSelect')
 MULTI_CHOICE_FIELDS = getattr(settings, "DATAFORMS_MULTI_CHOICE_FIELDS", ()) + ('SelectMultiple', 'CheckboxSelectMultiple')
 # These fields tie into the Choice Model
-CHOICE_FIELDS = getattr(settings, "DATAFORMS_CHOICE_FIELDS", ()) + (SINGLE_CHOICE_FIELDS + MULTI_CHOICE_FIELDS)
+CHOICE_FIELDS = (SINGLE_CHOICE_FIELDS + MULTI_CHOICE_FIELDS)
 # These fields are saved as Comma Delimited Strings (usefull for numbers)
 STATIC_CHOICE_FIELDS = getattr(settings, "DATAFORMS_STATIC_CHOICE_FIELDS", ())
