+++ docs/source/diagram.rst
@@ -1,0 +1,4 @@
+Model Diagram
+=============
+
+.. image:: dataforms_models.png
