--- docs/source/index.txt
@@ -1,21 +1,0 @@
-.. django-dataforms documentation master file
-
-django-dataforms documentation
-==============================
-
-See the API documentation below, or the `Getting Started guide <http://code.google.com/p/django-dataforms/wiki/GettingStarted>`_ to jump in.
-
-Contents:
-
-.. toctree::
-   :maxdepth: 2
-   
-   forms
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
